// src/app/core/enums/TypeState.ts
export enum TypeState {
  PENDIENTE = 1,
  SINCRONIZADO = 2,
  EN_EDICION = 3,
  GUARDADO = 4,
  EN_SINCRONIZACION = 5,
  RECHAZADA = 6,
  ERROR = 7,
  GUARDADO_PARCIAL = 8,
  APROBADA_SINCRONIZADO = 9,
  PENDIENTE_APROBACION_SINCRONIZADO = 10,
  REPROGRAMADA = 11,
  REPROGRAMADA_ENVIADA = 12,
  COMPROMETIDA = 13,
  EN_EJECUCION = 14,
}